+++ BusinessManagementApp/ViewModels/DetailsVMs/ProductCategoryDetailsVM.cs
@@ -1,6 +1,8 @@
 ﻿using BusinessManagementApp.Data;
 using BusinessManagementApp.Data.Model;
+using BusinessManagementApp.ViewModels.BusyIndicator;
 using BusinessManagementApp.ViewModels.Navigation;
+using BusinessManagementApp.Views.Dialogs;
 using CommunityToolkit.Mvvm.Input;
 using System;
 using System.ComponentModel.DataAnnotations;
@@ -99,7 +101,7 @@ namespace BusinessManagementApp.ViewModels.DetailsVMs
 
 
             Save = new AsyncRelayCommand(SaveProductCategory);
-            Delete = new AsyncRelayCommand(DeleteSkillType);
+            Delete = new RelayCommand(DeleteProductCategory);
             Cancel = new RelayCommand(
                 () => WorkspaceNavUtils.NavigateTo(WorkspaceViewName.ProductCategories)
                 );
@@ -110,7 +112,7 @@ namespace BusinessManagementApp.ViewModels.DetailsVMs
         public override async void LoadData(object? id = null)
         {
 
-
+            BusyIndicatorUtils.SetBusyIndicator(true);
             if (id != null)
             {
                 IsEditMode = true;
@@ -118,6 +120,7 @@ namespace BusinessManagementApp.ViewModels.DetailsVMs
             }
 
             CanSave = true;
+            BusyIndicatorUtils.SetBusyIndicator(false);
         }
 
         private async Task LoadProductCategory(int id)
@@ -130,6 +133,7 @@ namespace BusinessManagementApp.ViewModels.DetailsVMs
 
         private async Task SaveProductCategory()
         {
+            BusyIndicatorUtils.SetBusyIndicator(true);
             ValidateAllProperties();
             if (HasErrors)
                 return;
@@ -149,15 +153,28 @@ namespace BusinessManagementApp.ViewModels.DetailsVMs
             {
                 await productCategoriesRepo.AddProductCategory(productCategory);
             }
-
+            BusyIndicatorUtils.SetBusyIndicator(false);
             // Navigate back to list screen
-            WorkspaceNavUtils.NavigateTo(WorkspaceViewName.Providers);
+            WorkspaceNavUtils.NavigateTo(WorkspaceViewName.ProductCategories);
         }
 
-        private async Task DeleteSkillType()
+        private void DeleteProductCategory()
         {
-            await productCategoriesRepo.DeleteProductCategory(Id);
-            WorkspaceNavUtils.NavigateTo(WorkspaceViewName.ProductCategories);
+            ConfirmDialog dialog = new ConfirmDialog(
+            "Delete product category",
+            "Do you want to delete this product category?\n" +
+            "This action cannot be undone!");
+            dialog.Closed += async (sender, eventArgs) =>
+            {
+                if (dialog.IsConfirmed)
+                {
+                    BusyIndicatorUtils.SetBusyIndicator(true);
+                    await productCategoriesRepo.DeleteProductCategory(Id);
+                    BusyIndicatorUtils.SetBusyIndicator(false);
+                    WorkspaceNavUtils.NavigateTo(WorkspaceViewName.ProductCategories);
+                }
+            };
+            dialog.Show();
         }
     }
 }
